+++ src/lib/common/Events.ts
@@ -172,6 +172,7 @@ class Emitter extends EventEmitterClass{
  * @fires Events#rainbow_onbubblepollterminated
  * @fires Events#rainbow_onbubblepollupdated
  * @fires Events#rainbow_onbubblepollvoted
+ * @fires Events#rainbow_onbubbleroompasswordreceived
  * @fires Events#rainbow_onconnectorcommand
  * @fires Events#rainbow_onconnectorconfig
  * @fires Events#rainbow_onconnectorcommandended
@@ -305,6 +306,7 @@ class Events {
         "rainbow_onbubblepollterminated",
         "rainbow_onbubblepollupdated",
         "rainbow_onbubblepollvoted",
+        "rainbow_onbubbleroompasswordreceived",
         "rainbow_onconnectorcommand",
         "rainbow_onconnectorconfig",
         "rainbow_onconnectorcommandended",
@@ -1530,6 +1532,17 @@ class Events {
              *      This event is fired when a poll is voted in bubble.
              */
             that.publishEvent("bubblepollvoted", data);
+        });
+
+        this._evReceiver.on("evt_internal_bubble_roompassword_received", function (data) {
+            /**
+             * @event Events#rainbow_onbubbleroompasswordreceived
+             * @public
+             * @param { Object } data informations about poll
+             * @description
+             *      This event is fired when a password is setted for public links in a bubble.
+             */
+            that.publishEvent("bubbleroompasswordreceived", data);
         });
 
         this._evReceiver.on("evt_internal_connectorcommand", function (data) {