+++ src/lib/connection/XMPPServiceHandler/conversationEventHandler.ts
@@ -2300,6 +2300,7 @@ class ConversationEventHandler extends GenericHandler {
             let children = stanza.children;
             children.forEach(function (node) {
                 let nodeJson = jsonStanza.message[node.getName()];
+                that._logger.log(that.INTERNAL, LOG_ID + "(onManagementMessageReceived) nodeJson : ", nodeJson);
                 switch (node.getName()) {
                     case "room":
                         that.onRoomManagementMessageReceived(node, nodeJson);
@@ -2326,6 +2327,10 @@ class ConversationEventHandler extends GenericHandler {
                     case "vscan":
                         //treated in conversationEventHandler::onVscanManagementMessageReceived(node);
                         that.onVscanManagementMessageReceived(nodeJson);
+                        break;
+                    case "roompassword":
+                        //treated in conversationEventHandler::onRoomPasswordManagementMessageReceived(node);
+                        that.onRoomPasswordManagementMessageReceived(nodeJson);
                         break;
                     case "file":
                         that.onFileManagementMessageReceived(node);
@@ -2902,6 +2907,20 @@ class ConversationEventHandler extends GenericHandler {
             }
         } catch (err) {
             that._logger.log(that.ERROR, LOG_ID + "(onVscanManagementMessageReceived) CATCH Error !!! : ", err);
+        }
+    };
+
+    onRoomPasswordManagementMessageReceived (nodeJson) {
+        let that = this;
+        try {
+            that._logger.log(that.INTERNAL, LOG_ID + "(onRoomPasswordManagementMessageReceived) _entering_ : ", nodeJson);
+            if (nodeJson && nodeJson["$attrs"]) {
+                let nodeAttrs = Object.assign({}, nodeJson["$attrs"]);
+                delete nodeAttrs.xmlns;
+                that.eventEmitter.emit("evt_internal_bubble_roompassword_received", nodeAttrs);
+            }
+        } catch (err) {
+            that._logger.log(that.ERROR, LOG_ID + "(onRoomPasswordManagementMessageReceived) CATCH Error !!! : ", err);
         }
     };